requirements.txt
einops==0.8.1
lightning==2.4.0
pytorch_ignite==0.5.1
addict==2.4.0
matplotlib==3.10.1
numpy==1.26.3
open3d==0.18.0
opt_einsum==3.3.0
pandas==2.2.3
pykdtree==1.3.13
spconv_cu118==2.3.6
timm==1.0.15
torchmetrics==1.4.2
tqdm==4.66.5
jsonargparse[signatures]>=4.27.7
tensorboardX==2.6.2.2
pyarrow==19.0.1
